Add support for numbers
The party parrot text converter is the greatest piece of code gifted to mankind in the
last 20 years. But, it currently only supports the 26 letters in the English alphabet.
To rectify this, this PR adds in the ability to spice up life with numbers!1!

(also raises an exception if a character isn't supported)